refactor(tests): split narrativeGenerator and characterStore tests
Split narrativeGenerator.decision-consequences.test.ts (613 lines) into:
- integration.test.ts (154 lines) - Past decision integration tests
- mapping.test.ts (117 lines) - Decision consequence mapping
- longterm.test.ts (126 lines) - Long-term consequence building
- errors.test.ts (148 lines) - Error handling and edge cases
- decisionConsequences.testHelpers.ts (146 lines) - Shared test utilities

Split characterStore.test.ts (541 lines) into:
- crud.test.ts (181 lines) - Create, update, delete, setCurrentCharacter operations
- attributes.test.ts (145 lines) - Attribute and skill management
- state.test.ts (83 lines) - Initialization, error handling, loading state, reset
- testHelpers.ts (83 lines) - Reusable test data factories

Benefits:
- All test files now under 300 lines (complies with project standards)
- Eliminated duplicate test data setup through shared helpers
- Improved test discoverability by logical grouping
- Easier to locate and modify specific test scenarios

Progress: 3 of 48 files refactored (6%)
