--- src/lib/ai/__tests__/narrativeGenerator.decision-consequences.test.ts
@@ -1,613 +1,0 @@
-/**
- * TDD Tests for Issue #210: Make player decisions have meaningful story consequences
- * 
- * These tests verify that the AI narrative generation system actually uses past
- * player decisions to influence story content, not just collects them.
- */
-
-import { NarrativeGenerator } from '../narrativeGenerator';
-import { playerDecisionTracker } from '../playerDecisionTracker';
-import { useWorldStore } from '@/state/worldStore';
-import { useCharacterStore } from '@/state/characterStore';
-import { useAiContextStore } from '@/state/aiContextStore';
-import { MockAIClient } from '../../__mocks__/mockAiClient';
-import { NarrativeGenerationRequest } from '@/types/narrative.types';
-import { PlayerDecision } from '@/types/personalization.types';
-import { narrativeTemplateManager } from '../../promptTemplates/narrativeTemplateManager';
-import { getLoreContextForPrompt } from '../loreContextHelper';
-import { getDetailedToneInstructions } from '../toneSettingsGuidance';
-import { getTimestamp } from '@/lib/utils/timestamp';
-
-// Mock the store modules
-jest.mock('@/state/worldStore', () => ({
-  useWorldStore: {
-    getState: jest.fn()
-  }
-}));
-jest.mock('@/state/characterStore', () => ({
-  useCharacterStore: {
-    getState: jest.fn()
-  }
-}));
-jest.mock('@/state/aiContextStore', () => ({
-  useAiContextStore: {
-    getState: jest.fn()
-  }
-}));
-jest.mock('../playerDecisionTracker');
-jest.mock('../../promptTemplates/narrativeTemplateManager', () => ({
-  narrativeTemplateManager: {
-    getTemplate: jest.fn()
-  }
-}));
-jest.mock('../loreContextHelper', () => ({
-  getLoreContextForPrompt: jest.fn()
-}));
-jest.mock('../structuredLoreExtractor', () => ({
-  extractStructuredLore: jest.fn()
-}));
-jest.mock('../toneSettingsGuidance', () => ({
-  getDetailedToneInstructions: jest.fn()
-}));
-jest.mock('@/state/loreStore', () => ({
-  useLoreStore: {
-    getState: jest.fn()
-  }
-}));
-
-const mockPlayerDecisionTracker = playerDecisionTracker as jest.Mocked<typeof playerDecisionTracker>;
-
-describe('NarrativeGenerator - Decision Consequences (Issue #210)', () => {
-  let narrativeGenerator: NarrativeGenerator;
-  let mockAiClient: MockAIClient;
-  let pastDecisions: PlayerDecision[];
-
-  const mockWorld = {
-    id: 'world-1',
-    name: 'Test World',
-    theme: 'Fantasy Adventure',
-    description: 'A magical realm',
-    attributes: [],
-    skills: [],
-    settings: { maxAttributes: 6, maxSkills: 10, attributePointPool: 30, skillPointPool: 50 }
-  };
-
-  const mockCharacter = {
-    id: 'char-1',
-    name: 'Hero',
-    worldId: 'world-1',
-    background: { description: 'A brave warrior', personality: 'Courageous', motivation: 'Save the kingdom' },
-    attributes: [],
-    skills: [],
-    createdAt: new Date(),
-    updatedAt: new Date()
-  };
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-    // Use fake timers for deterministic timestamp generation
-    jest.useFakeTimers();
-    jest.setSystemTime(new Date('2025-01-15T12:00:00Z'));
-
-    // Create past decisions with deterministic timestamps
-    // Set time to 1 day ago for first decision
-    jest.setSystemTime(new Date('2025-01-14T12:00:00Z'));
-    const oneDayAgo = getTimestamp();
-
-    // Set time to 12 hours ago for second decision
-    jest.setSystemTime(new Date('2025-01-15T00:00:00Z'));
-    const twelveHoursAgo = getTimestamp();
-
-    // Reset to "now"
-    jest.setSystemTime(new Date('2025-01-15T12:00:00Z'));
-
-    pastDecisions = [
-      {
-        id: 'decision-1',
-        prompt: 'An injured merchant lies on the ground. What do you do?',
-        sessionId: 'session-1',
-        worldId: 'world-1',
-        choiceText: 'Help the injured merchant',
-        choiceType: 'helpful',
-        timestamp: oneDayAgo,
-        context: { situation: 'moral dilemma', charactersPresent: ['merchant'] }
-      },
-      {
-        id: 'decision-2',
-        prompt: 'Bandits block your path. How do you handle this?',
-        sessionId: 'session-1',
-        worldId: 'world-1',
-        choiceText: 'Negotiate instead of fighting',
-        choiceType: 'diplomatic',
-        timestamp: twelveHoursAgo,
-        context: { situation: 'conflict resolution', charactersPresent: ['bandit leader'] }
-      }
-    ];
-    mockAiClient = new MockAIClient();
-    narrativeGenerator = new NarrativeGenerator(mockAiClient);
-
-    // Mock world store directly
-    (useWorldStore.getState as jest.Mock).mockReturnValue({
-      worlds: { 'world-1': mockWorld },
-      currentWorldId: 'world-1',
-      error: null,
-      loading: false
-    });
-
-    // Mock character store directly
-    (useCharacterStore.getState as jest.Mock).mockReturnValue({
-      characters: { 'char-1': mockCharacter },
-      currentCharacterId: 'char-1',
-      error: null,
-      loading: false
-    });
-
-    // Mock aiContext store for goal context
-    (useAiContextStore as unknown as { getState: jest.Mock }).getState.mockReturnValue({
-      buildContextForSession: jest.fn().mockReturnValue({
-        goalContext: null,
-        activeGoals: []
-      })
-    });
-
-    mockPlayerDecisionTracker.getWorldDecisions.mockReturnValue(pastDecisions);
-    mockPlayerDecisionTracker.getRelevantDecisions.mockReturnValue(pastDecisions);
-
-    // Mock getRelevantDecisionsWithScores to return decisions with high relevance scores
-    mockPlayerDecisionTracker.getRelevantDecisionsWithScores.mockReturnValue(
-      pastDecisions.map(decision => ({
-        decision,
-        relevanceScore: {
-          decisionId: decision.id,
-          overallScore: 0.8, // High relevance to ensure full detail formatting
-          recencyScore: 0.8,
-          contextScore: 0.8,
-          impactScore: 0.8,
-          tagMatchScore: 0.8,
-          characterScore: 0.8,
-          calculatedAt: getTimestamp()
-        }
-      }))
-    );
-
-    // Mock template manager to return a simple template function
-    (narrativeTemplateManager.getTemplate as jest.Mock).mockReturnValue(
-      (context: { worldName: string; characterIds?: string[] }) => `Generated prompt for ${context.worldName} with ${context.characterIds?.length || 0} characters`
-    );
-
-    // Mock lore context helper
-    (getLoreContextForPrompt as jest.Mock).mockReturnValue('');
-
-    // Mock tone settings guidance
-    (getDetailedToneInstructions as jest.Mock).mockReturnValue('');
-  });
-
-  afterEach(() => {
-    jest.useRealTimers();
-  });
-
-  describe('Past Decision Integration in Prompts', () => {
-    test('should include specific past decisions in AI prompts', async () => {
-      // ACCEPTANCE CRITERIA: AI prompts must include specific past decision details
-      const request: NarrativeGenerationRequest = {
-        worldId: 'world-1',
-        characterIds: ['char-1'],
-        sessionId: 'session-1'
-      };
-
-      await narrativeGenerator.generateSegment(request);
-
-            // Verify that the AI prompt includes specific decision context
-
-            const narrativePrompt = mockAiClient.getPrompts()[0];
-
-            
-
-            // Should mention specific past choices (new format: lowercase, may be partial)
-            // New format: "- At [location], you [action] (type)"
-            const hasHelpMerchant = narrativePrompt.toLowerCase().includes('help') &&
-                                   narrativePrompt.toLowerCase().includes('merchant');
-            const hasNegotiate = narrativePrompt.toLowerCase().includes('negotiate');
-
-            expect(hasHelpMerchant).toBe(true);
-            expect(hasNegotiate).toBe(true);
-
-            
-
-            // Should include decision context section
-
-            expect(narrativePrompt).toMatch(/RECENT PLAYER DECISIONS?:/i);
-
-            
-
-            // Should include any contextual information from decisions (flexible)
-
-            const hasContextualInfo = narrativePrompt.includes('moral dilemma') || 
-
-                                     narrativePrompt.includes('merchant') || 
-
-                                     narrativePrompt.includes('compassionate') ||
-
-                                     narrativePrompt.includes('diplomatic');
-
-            expect(hasContextualInfo).toBe(true);
-
-          });
-
-      
-
-          test('should instruct AI to reference past decisions in narrative', async () => {
-
-            // ACCEPTANCE CRITERIA: AI must be explicitly instructed to use past decisions
-
-            const request: NarrativeGenerationRequest = {
-
-              worldId: 'world-1',
-
-              characterIds: ['char-1'],
-
-              sessionId: 'session-1'
-
-            };
-
-      
-
-            await narrativeGenerator.generateSegment(request);
-
-      
-
-            const narrativePrompt = mockAiClient.getPrompts()[0];
-
-            
-
-            // Decision history provides context for the AI to use
-            // The presence of "RECENT PLAYER DECISIONS" section serves as implicit instruction
-            const hasDecisionSection = narrativePrompt.includes('RECENT PLAYER DECISIONS');
-            expect(hasDecisionSection).toBe(true);
-
-            // Verify decisions are actually included (not just the header)
-            const decisionLines = narrativePrompt.split('\n').filter(line => line.trim().startsWith('- At'));
-            expect(decisionLines.length).toBeGreaterThan(0);
-
-          });
-
-      
-
-          test('should include decision impact instructions for NPCs', async () => {
-
-            // ACCEPTANCE CRITERIA: NPCs should react based on past player interactions
-
-            const request: NarrativeGenerationRequest = {
-
-              worldId: 'world-1', 
-
-              characterIds: ['char-1'],
-
-              sessionId: 'session-1'
-
-            };
-
-      
-
-            await narrativeGenerator.generateSegment(request);
-
-      
-
-            const narrativePrompt = mockAiClient.getPrompts()[0];
-
-            
-
-            // Check that decisions involving NPCs are included
-            const promptLower = narrativePrompt.toLowerCase();
-            const mentionsMerchant = promptLower.includes('merchant');
-            expect(mentionsMerchant).toBe(true);
-
-            // Check that the decision history includes relevant choices
-            const hasHelpReference = promptLower.includes('help') ||
-                                    promptLower.includes('assist');
-            expect(hasHelpReference).toBe(true);
-
-          });
-
-        });
-
-      
-
-        describe('Decision Consequence Mapping', () => {
-
-          test('should map compassionate decisions to trust-building consequences', async () => {
-
-            // ACCEPTANCE CRITERIA: Compassionate choices should build trust and reputation
-
-            const request: NarrativeGenerationRequest = {
-
-              worldId: 'world-1',
-
-              characterIds: ['char-1'], 
-
-              sessionId: 'session-1'
-
-            };
-
-      
-
-            await narrativeGenerator.generateSegment(request);
-
-      
-
-            const narrativePrompt = mockAiClient.getPrompts()[0];
-
-            
-
-            // Check that decision data is included (new format uses lowercase and (type) not [type])
-            const promptLower = narrativePrompt.toLowerCase();
-            const hasHelpDecision = promptLower.includes('help') && promptLower.includes('merchant');
-            const hasHelpfulType = promptLower.includes('(helpful)');
-
-            expect(hasHelpDecision).toBe(true);
-            expect(hasHelpfulType).toBe(true);
-
-          });
-
-      
-
-          test('should map diplomatic decisions to peaceful resolution options', async () => {
-
-            // ACCEPTANCE CRITERIA: Diplomatic choices should create peaceful alternatives
-
-            const request: NarrativeGenerationRequest = {
-
-              worldId: 'world-1',
-
-              characterIds: ['char-1'],
-
-              sessionId: 'session-1'
-
-            };
-
-      
-
-            await narrativeGenerator.generateSegment(request);
-
-      
-
-            const narrativePrompt = mockAiClient.getPrompts()[0];
-
-            
-
-            // Check that decision data is included (LLM will infer diplomatic consequences)
-
-            // Check for diplomatic decision in new format
-            const promptLower = narrativePrompt.toLowerCase();
-            const hasNegotiateDecision = promptLower.includes('negotiate');
-            const hasDiplomaticType = promptLower.includes('(diplomatic)');
-
-            expect(hasNegotiateDecision).toBe(true);
-            expect(hasDiplomaticType).toBe(true);
-
-          });
-
-        });
-
-      
-
-        describe('Long-term Consequence Building', () => {
-
-          test('should create narrative threads that span multiple decisions', async () => {
-
-            // ACCEPTANCE CRITERIA: Multiple related decisions should build coherent consequences
-
-            const request: NarrativeGenerationRequest = {
-
-              worldId: 'world-1',
-
-              characterIds: ['char-1'],
-
-              sessionId: 'session-1'
-
-            };
-
-      
-
-            await narrativeGenerator.generateSegment(request);
-
-      
-
-            const narrativePrompt = mockAiClient.getPrompts()[0];
-
-            
-
-            // Check that multiple decisions are included (LLM will infer patterns)
-            const promptLower = narrativePrompt.toLowerCase();
-            const hasHelpDecision = promptLower.includes('help') && promptLower.includes('merchant');
-            const hasNegotiateDecision = promptLower.includes('negotiate');
-
-            expect(hasHelpDecision).toBe(true);
-            expect(hasNegotiateDecision).toBe(true);
-
-      
-
-            // Check that choice types are tagged for pattern recognition
-
-            const hasChoiceTypes = promptLower.includes('(helpful)') &&
-                                  promptLower.includes('(diplomatic)');
-
-            expect(hasChoiceTypes).toBe(true);
-
-          });
-
-      
-
-          test('should provide decision-informed story options', async () => {
-
-            // ACCEPTANCE CRITERIA: Narrative should suggest options that reflect past decision patterns
-
-            const request: NarrativeGenerationRequest = {
-
-              worldId: 'world-1',
-
-              characterIds: ['char-1'],
-
-              sessionId: 'session-1'
-
-            };
-
-      
-
-            await narrativeGenerator.generateSegment(request);
-
-      
-
-            const narrativePrompt = mockAiClient.getPrompts()[0];
-
-            
-
-            // The decision history section includes past choices
-            const hasDecisionSection = narrativePrompt.includes('RECENT PLAYER DECISIONS');
-            expect(hasDecisionSection).toBe(true);
-
-            // Should reference decision types (helpful, diplomatic, etc.)
-            const promptLower = narrativePrompt.toLowerCase();
-            const hasDecisionTypes = promptLower.includes('(helpful)') ||
-                                    promptLower.includes('(diplomatic)') ||
-                                    promptLower.includes('(aggressive)') ||
-                                    promptLower.includes('(chaotic)');
-            expect(hasDecisionTypes).toBe(true);
-
-          });
-
-        });
-
-      
-
-        describe('Error Handling and Edge Cases', () => {
-
-          test('should handle empty decision history gracefully', async () => {
-
-            // When no past decisions exist, should not break
-
-            mockPlayerDecisionTracker.getWorldDecisions.mockReturnValue([]);
-            mockPlayerDecisionTracker.getRelevantDecisions.mockReturnValue([]);
-            mockPlayerDecisionTracker.getRelevantDecisionsWithScores.mockReturnValue([]);
-
-            
-
-            const request: NarrativeGenerationRequest = {
-
-              worldId: 'world-1',
-
-              characterIds: ['char-1'],
-
-              sessionId: 'session-1'
-
-            };
-
-      
-
-            await expect(narrativeGenerator.generateSegment(request)).resolves.toBeDefined();
-
-          });
-
-      
-
-          test('should limit decision history to prevent prompt overflow', async () => {
-
-            // Create many past decisions with deterministic timestamps
-
-            const manyDecisions = Array.from({ length: 20 }, (_, i) => {
-
-              // Set time to i seconds ago
-
-              const timeAgo = new Date('2025-01-15T12:00:00Z');
-
-              timeAgo.setSeconds(timeAgo.getSeconds() - i);
-
-              jest.setSystemTime(timeAgo);
-
-              const timestamp = getTimestamp();
-
-      
-
-              return {
-
-                id: `decision-${i}`,
-
-                prompt: `What will you do next? (Decision ${i})`,
-
-                sessionId: 'session-1',
-
-                worldId: 'world-1',
-
-                choiceText: `Choice ${i}`,
-
-                choiceType: 'neutral' as const,
-
-                timestamp,
-
-                context: { situation: 'generic' }
-
-              };
-
-            });
-
-      
-
-            // Reset to "now"
-
-            jest.setSystemTime(new Date('2025-01-15T12:00:00Z'));
-
-      
-
-            mockPlayerDecisionTracker.getWorldDecisions.mockReturnValue(manyDecisions);
-            mockPlayerDecisionTracker.getRelevantDecisions.mockReturnValue(manyDecisions);
-            mockPlayerDecisionTracker.getRelevantDecisionsWithScores.mockReturnValue(
-              manyDecisions.map(decision => ({
-                decision,
-                relevanceScore: {
-                  decisionId: decision.id,
-                  overallScore: 0.5,
-                  recencyScore: 0.5,
-                  contextScore: 0.5,
-                  impactScore: 0.5,
-                  tagMatchScore: 0.5,
-                  characterScore: 0.5,
-                  calculatedAt: getTimestamp()
-                }
-              }))
-            );
-
-      
-
-            const request: NarrativeGenerationRequest = {
-
-              worldId: 'world-1',
-
-              characterIds: ['char-1'],
-
-              sessionId: 'session-1'
-
-            };
-
-      
-
-            await narrativeGenerator.generateSegment(request);
-
-      
-
-            const narrativePrompt = mockAiClient.getPrompts()[0];
-
-            
-
-            // Should include decisions but not overwhelm the prompt
-
-            const decisionMatches = narrativePrompt.match(/Choice \d+/g) || [];
-
-            expect(decisionMatches.length).toBeLessThanOrEqual(10); // Reasonable limit
-
-          });
-
-        });
-
-      });
-
-
