+++ src/state/__tests__/characterStore.state.test.ts
@@ -1,0 +1,83 @@
+/**
+ * Tests for CharacterStore State Management
+ *
+ * Verifies initialization, error handling, loading state, and reset functionality.
+ */
+
+import { useCharacterStore } from '../characterStore';
+import { ErrorType } from '@/lib/utils/errorUtils';
+import { createTestCharacterData, setupTestTimers } from './characterStore.testHelpers';
+
+describe('useCharacterStore - State Management', () => {
+  beforeEach(() => {
+    setupTestTimers();
+    useCharacterStore.getState().reset();
+  });
+
+  afterEach(() => {
+    jest.clearAllTimers();
+    jest.useRealTimers();
+  });
+
+  describe('initialization', () => {
+    test('should initialize with default state', () => {
+      const state = useCharacterStore.getState();
+      expect(state.characters).toEqual({});
+      expect(state.entities).toEqual({});
+      expect(state.currentCharacterId).toBeNull();
+      expect(state.currentEntityId).toBeNull();
+      expect(state.error).toBeNull();
+      expect(state.loading).toBe(false);
+    });
+  });
+
+  describe('error handling', () => {
+    test('should set and clear errors', () => {
+      useCharacterStore.getState().setError({
+        title: 'Test error',
+        message: 'Details',
+        retryable: false,
+        type: ErrorType.UNKNOWN
+      });
+      expect(useCharacterStore.getState().error?.title).toBe('Test error');
+
+      useCharacterStore.getState().clearError();
+      expect(useCharacterStore.getState().error).toBeNull();
+    });
+  });
+
+  describe('loading state', () => {
+    test('should set loading state', () => {
+      useCharacterStore.getState().setLoading(true);
+      expect(useCharacterStore.getState().loading).toBe(true);
+
+      useCharacterStore.getState().setLoading(false);
+      expect(useCharacterStore.getState().loading).toBe(false);
+    });
+  });
+
+  describe('reset', () => {
+    test('should reset store to initial state', () => {
+      // Add some data
+      useCharacterStore.getState().createCharacter(createTestCharacterData());
+      useCharacterStore.getState().setError({
+        title: 'Some error',
+        message: 'Details',
+        retryable: false,
+        type: ErrorType.UNKNOWN
+      });
+      useCharacterStore.getState().setLoading(true);
+
+      // Reset
+      useCharacterStore.getState().reset();
+      const state = useCharacterStore.getState();
+
+      expect(state.characters).toEqual({});
+      expect(state.entities).toEqual({});
+      expect(state.currentCharacterId).toBeNull();
+      expect(state.currentEntityId).toBeNull();
+      expect(state.error).toBeNull();
+      expect(state.loading).toBe(false);
+    });
+  });
+});
